+++ internal/reportdata/telegraf.go
@@ -5,6 +5,8 @@ import (
 	"net/http"
 	"strconv"
 	"time"
+
+	"github.com/egorkovalchuk/go-ltreport/internal/logger"
 )
 
 // TelegrafClient представляет клиент для работы
@@ -12,12 +14,12 @@ type TelegrafClient struct {
 	baseURL string
 	auth    string
 	client  *http.Client
-	logFunc func(string, interface{})
+	logFunc *logger.LogWriter
 	debug   bool
 }
 
 // NewTelegrafClient создает новый экземпляр клиента
-func NewTelegrafClient(baseURL string, auth string, logFunc func(string, interface{}), debug bool) *TelegrafClient {
+func NewTelegrafClient(baseURL string, auth string, logFunc *logger.LogWriter, debug bool) *TelegrafClient {
 	return &TelegrafClient{
 		baseURL: baseURL,
 		auth:    auth,
@@ -31,12 +33,6 @@ func (p *TelegrafClient) Close() {
 	p.client.CloseIdleConnections()
 }
 
-func (p *TelegrafClient) ProcessDebug(t interface{}) {
-	if p.debug {
-		p.logFunc("DEBUG", t)
-	}
-}
-
 func (p *TelegrafClient) GetDataSourceThreshold(query string) (float64, error) {
 	var percentile float64
 	resp_inf, err := http.NewRequest("GET", p.baseURL+""+query, nil)
@@ -45,7 +41,7 @@ func (p *TelegrafClient) GetDataSourceThreshold(query string) (float64, error) {
 	}
 	resp_inf.Header.Add("Authorization", p.auth)
 
-	p.ProcessDebug("Get telegraf threshold request: " + p.baseURL + "/api/v1/query?query=" + query)
+	p.logFunc.ProcessDebug("Get telegraf threshold request: " + p.baseURL + "/api/v1/query?query=" + query)
 	rsp_inf, err := p.client.Do(resp_inf)
 
 	if err != nil {
@@ -54,7 +50,7 @@ func (p *TelegrafClient) GetDataSourceThreshold(query string) (float64, error) {
 	defer rsp_inf.Body.Close()
 
 	if rsp_inf.StatusCode == http.StatusOK {
-		p.logFunc("INFO", "Request prometheus threshold success")
+		p.logFunc.ProcessInfo("Request prometheus threshold success")
 
 		var prom PrometheusResponse
 		err = prom.JsonPrometheusParse(rsp_inf)